[ft #161278931] added functionality for a user to like and dislike a comment
