+++ .gitignore
@@ -96,5 +96,4 @@ ENV/
 # SQLite3
 db.sqlite3
 .vscode/
-migrations/
 .DS_Store
