+++ authors/apps/articles/serializers.py
@@ -1,16 +1,13 @@
-"""
 
-Imports
-
-"""
-from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 from rest_framework.pagination import PageNumberPagination
+from rest_framework import serializers
 
 from authors.apps.articles.utils import ChoicesField
-from .models import Article, TaggedItem, ArticleRating, Photo, LikeDislike, Comment
+from .models import Article, TaggedItem, ArticleRating, Photo
 from authors.apps.profiles.serializers import ProfileSerializer
 from authors.apps.profiles.models import Profile
+from authors.apps.likedislike.serializers import VoteObjectRelatedSerializer
 
 class ArticlePagination(PageNumberPagination):
     """
@@ -39,58 +36,6 @@ class TaggedOjectRelatedField(serializers.RelatedField):
     def to_internal_value(self, data):
         return TaggedItem.objects.create(tag_name=data)
 
-class RecursiveSerializer(serializers.Serializer):
-    def to_representation(self, value):
-        serializer = self.parent.parent.__class__(value, context=self.context)
-        return serializer.data
-
-
-class CommentSerializer(serializers.ModelSerializer):
-    """This class creates the comments serializers"""
-
-    author = serializers.SerializerMethodField()
-    article = serializers.ReadOnlyField(source='article.title')
-    thread = RecursiveSerializer(many=True, read_only=True)
-
-
-    class Meta:
-        model = Comment
-
-        fields = (
-            'id',
-            'createdAt',
-            'updatedAt',
-            'body',
-            'article',
-            'author',
-            'thread'
-        )
-
-
-    def get_author(self, obj):
-        try:
-            author = obj.author
-            profile = Profile.objects.get(user_id=author.id)
-            serializer = ProfileSerializer(profile)
-            return serializer.data
-        except Exception as e:
-            return {}
-
-    def create(self, validated_data):
-        """
-        Create and return a new comment instance, given the validated_data
-        """
-        parent = self.context.get('parent', None)
-        instance = Comment.objects.create(parent=parent, **validated_data)
-        return instance
-
-    def update(self, instance, validated_data, **kwargs):
-        """
-        Update and return a comment instance, given the validated_data
-        """
-        instance.body = validated_data.get('body', instance.body)
-        instance.save()
-        return instance
 
 
 class ArticleSerializer(serializers.ModelSerializer):
@@ -99,7 +44,7 @@ class ArticleSerializer(serializers.ModelSerializer):
     """
     author = serializers.ReadOnlyField(source='author.username')
     tags = TaggedOjectRelatedField(many=True)
-
+    votes = VoteObjectRelatedSerializer(read_only=True)
     class Meta:
         model = Article
 
@@ -111,9 +56,11 @@ class ArticleSerializer(serializers.ModelSerializer):
             'createdAt',
             'updatedAt',
             'favorited',
+            'votes',
             'body',
             'read_time',
-            'author'
+            'author',
+
         )
 
 
@@ -179,12 +126,3 @@ class RatingSerializer(serializers.ModelSerializer):
                                   "Consider updating your rating")
         instance = ArticleRating.objects.create(**validated_data)
         return instance
-
-class LikeDislikeSerializer(serializers.Serializer):
-    """
-    serializer class for the Like Dislike model
-    """
-    class Meta:
-        model = LikeDislike
-        pass
-
