+++ authors/apps/articles/urls.py
@@ -2,9 +2,12 @@
  All Imports request on the file
 """
 
-from django.urls import path
-from .views import ArticleAPIView, ArticleRetrieveAPIView, LikeDislikeView, RateAPIView, RateRetrieveAPIView, ArticleCommentAPIView, ArticleCommentUpdateDeleteAPIView
-from .models import Article, LikeDislike
+from django.urls import path, include
+from .views import (ArticleAPIView, ArticleRetrieveAPIView, 
+                    RateAPIView, RateRetrieveAPIView )
+from .models import Article
+from authors.apps.likedislike.views import LikeDislikeView
+from authors.apps.likedislike.models import LikeDislike
 
 app_name = 'articles'
 
@@ -13,12 +16,10 @@ urlpatterns = [
     path('<slug:slug>', ArticleRetrieveAPIView.as_view(), name="retrieve_article"),
     path('<slug:slug>/rate', RateAPIView.as_view(), name="create_get_rating"),
     path('rate/detail/<int:rate_id>', RateRetrieveAPIView.as_view(), name="get_update_delete_rating"),
-    path('<slug:slug>/like/',
+    path('<slug:slug>/like',
          LikeDislikeView.as_view(model=Article, vote_type=LikeDislike.LIKE),
          name='article_like'),
-    path('<slug:slug>/dislike/',
+    path('<slug:slug>/dislike',
          LikeDislikeView.as_view(model=Article, vote_type=LikeDislike.DISLIKE),
-         name='article_dislike'),
-    path('<slug:slug>/comments/', ArticleCommentAPIView.as_view(), name='comment_article'),
-    path('<slug:slug>/comments/<int:id>/', ArticleCommentUpdateDeleteAPIView.as_view(), name='update_comment')
+         name='article_dislike')
 ]
