+++ authors/apps/comments/admin.py
@@ -1,0 +1,9 @@
+"""
+Imports
+"""
+from django.contrib import admin
+
+from .models import Comment
+
+
+admin.site.register(Comment)
