+++ authors/apps/comments/tests/test_comment_on_article.py
@@ -12,10 +12,8 @@ class TestComments(BaseTestCase):
         """test all fields required to comment on an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
-        data = {
-
-        }
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
+        data = {}
         response2 = self.client.post(url, data)
         self.assertEqual(response2.status_code, 400)
         assert(response2.data['body'][0] == "This field is required.")
@@ -24,7 +22,7 @@ class TestComments(BaseTestCase):
         """A user should be able to add a comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
         self.assertEqual(response2.status_code, 201)
 
@@ -33,7 +31,7 @@ class TestComments(BaseTestCase):
         """A user should be able to add a comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': 'slug'})
+        url = reverse('comments:comment_article', kwargs={'slug': 'slug'})
         response2 = self.client.post(url, self.new_comment)
         self.assertEqual(response2.status_code, 404)
         assert(response2.data['detail'] == "Not found.")
@@ -43,7 +41,7 @@ class TestComments(BaseTestCase):
         """A user should be able to get comments of an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.get(url)
         self.assertEqual(response2.status_code, 200)
         
@@ -51,12 +49,12 @@ class TestComments(BaseTestCase):
         """A user should be able to edit a comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
         data = {
             "body":"new comment"
         }
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         response3 = self.client.put(url, data)
         self.assertEqual(response3.status_code, 201)
 
@@ -65,9 +63,9 @@ class TestComments(BaseTestCase):
         """A user should be able to delete a comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         response3 = self.client.delete(url)
         self.assertEqual(response3.status_code, 200)
 
@@ -75,7 +73,7 @@ class TestComments(BaseTestCase):
         """A user should not be able to add an empty comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         data = {
             "body":""
         }
@@ -86,12 +84,12 @@ class TestComments(BaseTestCase):
         """A user can only edit their comments"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
         data = {
             "body":"new comment"
         }
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token2)
         response = self.client.put(url, data)
         self.assertEqual(response.status_code, 403)
@@ -100,9 +98,9 @@ class TestComments(BaseTestCase):
         """A user can only edit their comments"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token2)
         response = self.client.delete(url)
         self.assertEqual(response.status_code, 403)
@@ -111,12 +109,12 @@ class TestComments(BaseTestCase):
         """A user should be able to edit a comment to an article"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
         data = {
             "body":"new comment"
         }
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':1000})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':1000})
         response3 = self.client.put(url, data)
         self.assertEqual(response3.status_code, 404)
 
@@ -124,12 +122,12 @@ class TestComments(BaseTestCase):
         """A user should be able to thread a comment"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
         data = {
             "body":"new comment"
         }
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         response3 = self.client.post(url, data)
         self.assertEqual(response3.status_code, 201)
 
@@ -137,9 +135,9 @@ class TestComments(BaseTestCase):
         """test that a user can get a comment with its threads"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':response2.data['id']})
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token2)
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
@@ -148,9 +146,9 @@ class TestComments(BaseTestCase):
         """test that a user can get a comment with its threads"""
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
         response = self.client.post(self.article_url, self.new_article)
-        url = reverse('articles:comment_article', kwargs={'slug': response.data['slug']})
+        url = reverse('comments:comment_article', kwargs={'slug': response.data['slug']})
         response2 = self.client.post(url, self.new_comment)
-        url = reverse('articles:update_comment', kwargs={'slug': response.data['slug'], 'id':5})
+        url = reverse('comments:update_comment', kwargs={'slug': response.data['slug'], 'id':5})
         self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token2)
         response = self.client.get(url)
         self.assertEqual(response.status_code, 404)
