+++ authors/apps/likedislike/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import LikeDislike
+# Register your models here.
+
+admin.site.register(LikeDislike)
