+++ authors/apps/likedislike/urls.py
@@ -1,0 +1,5 @@
+from django.urls import path
+
+app_name = 'likedislike'
+
+url_patterns = []
