+++ authors/settings.py
@@ -12,6 +12,8 @@ https://docs.djangoproject.com/en/1.11/ref/settings/
 
 import os
 
+import datetime
+
 import django_heroku
 
 import dj_database_url
@@ -27,7 +29,7 @@ SECRET_KEY = os.getenv("SECRET_KEY") or "secret"
 
 APP_SETTINGS = os.getenv("APP_SETTINGS")
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True if APP_SETTINGS == "development" else False
+DEBUG = True #if APP_SETTINGS == "development" else False
 
 ALLOWED_HOSTS = ['*']
 # Application definition
@@ -60,7 +62,9 @@ INSTALLED_APPS = [
     'authors.apps.core',
     'authors.apps.profiles',
     'authors.apps.articles',
-    'authors.apps.friends'
+    'authors.apps.friends',
+    'authors.apps.likedislike',
+    'authors.apps.comments'
 ]
 
 MIDDLEWARE = [
@@ -231,15 +235,12 @@ SWAGGER_SETTINGS = {
 REST_AUTH_SERIALIZERS = {
     "USER_DETAILS_SERIALIZER":'authors.apps.authentication.serializers.CustomUserDetailsSerializer'
 }
+
 django_heroku.settings(locals(), logging=not DEBUG)
 
-locals()['DATABASES']['default'] = dj_database_url.config(default=os.getenv("DATABASE_URL"), ssl_require=not DEBUG)
-
-SECRET_KEY = os.getenv('SECRET_KEY')
-import datetime
+locals()['DATABASES']['default'] = dj_database_url.config(engine='django.db.backends.postgresql',
+                                                          ssl_require=not DEBUG)
 
 TOKEN_EXPIRE_TIME = datetime.timedelta(seconds=3)
 
-
-SECRET_KEY = os.getenv('SECRET_KEY')
